requirements.txt
httpx>=0.27,<1
google-genai>=1,<2
pytest>=8,<9
respx>=0.21,<1
